# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API server for the Omnichannel Chat Platform.",
        "title": "Omnichannel Chat Platform API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/contacts": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a list of customer contacts",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contacts"
                ],
                "summary": "List contacts",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search by name, email, phone",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (default 20)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add a new customer contact",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contacts"
                ],
                "summary": "Create contact",
                "parameters": [
                    {
                        "description": "Contact Info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateContactInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Contact"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/contacts/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get details of a specific contact",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contacts"
                ],
                "summary": "Get contact",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Contact ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Contact"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remove a contact",
                "tags": [
                    "contacts"
                ],
                "summary": "Delete contact",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Contact ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update contact details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contacts"
                ],
                "summary": "Update contact",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Contact ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Contact Info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateContactInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Contact"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/contacts/{id}/conversations": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get history of conversations with a contact",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contacts"
                ],
                "summary": "Get contact conversations",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Contact ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/models.Conversation"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.ChannelPublic": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.ChannelStatus"
                },
                "type": {
                    "$ref": "#/definitions/models.ChannelType"
                }
            }
        },
        "models.ChannelStatus": {
            "type": "string",
            "enum": [
                "active",
                "disconnected",
                "pending"
            ],
            "x-enum-varnames": [
                "ChannelStatusActive",
                "ChannelStatusDisconnected",
                "ChannelStatusPending"
            ]
        },
        "models.ChannelType": {
            "type": "string",
            "enum": [
                "whatsapp",
                "instagram",
                "facebook"
            ],
            "x-enum-varnames": [
                "ChannelWhatsApp",
                "ChannelInstagram",
                "ChannelFacebook"
            ]
        },
...
        "models.Contact": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "facebook_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "instagram_id": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object"
                },
                "name": {
                    "type": "string"
                },
                "organization_id": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "whatsapp_id": {
                    "description": "External IDs for different channels",
                    "type": "string"
                }
            }
        },
        "models.Conversation": {
            "type": "object",
            "properties": {
                "assigned_to": {
                    "type": "string"
                },
                "assigned_user": {
                    "$ref": "#/definitions/models.UserPublic"
                },
                "channel": {
                    "description": "Relations (populated when needed)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.ChannelPublic"
                        }
                    ]
                },
                "channel_id": {
                    "type": "string"
                },
                "contact": {
                    "$ref": "#/definitions/models.Contact"
                },
                "contact_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_human_reply_at": {
                    "description": "Added for AI Hybrid mode",
                    "type": "string"
                },
                "last_message": {
                    "$ref": "#/definitions/models.Message"
                },
                "last_message_at": {
                    "type": "string"
                },
                "organization_id": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.ConversationStatus"
                },
                "subject": {
                    "type": "string"
                },
                "unread_count": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.ConversationStatus": {
            "type": "string",
            "enum": [
                "open",
                "pending",
                "resolved",
                "closed"
            ],
            "x-enum-varnames": [
                "ConversationStatusOpen",
                "ConversationStatusPending",
                "ConversationStatusResolved",
                "ConversationStatusClosed"
            ]
        },
...
        "models.CreateContactInput": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "facebook_id": {
                    "type": "string"
                },
                "instagram_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 200,
                    "minLength": 1
                },
                "phone": {
                    "type": "string"
                },
                "whatsapp_id": {
                    "type": "string"
                }
            }
        },
...
        "models.Message": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "conversation_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "external_id": {
                    "description": "WhatsApp/IG message ID",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "media_file_name": {
                    "type": "string"
                },
                "media_mime_type": {
                    "type": "string"
                },
                "media_url": {
                    "type": "string"
                },
                "message_type": {
                    "$ref": "#/definitions/models.MessageType"
                },
                "reply_to": {
                    "$ref": "#/definitions/models.Message"
                },
                "reply_to_id": {
                    "type": "string"
                },
                "sender": {
                    "description": "Relations",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.UserPublic"
                        }
                    ]
                },
                "sender_id": {
                    "description": "Contact ID or User ID",
                    "type": "string"
                },
                "sender_type": {
                    "$ref": "#/definitions/models.SenderType"
                },
                "status": {
                    "$ref": "#/definitions/models.MessageStatus"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "models.MessageStatus": {
            "type": "string",
            "enum": [
                "pending",
                "sent",
                "delivered",
                "read",
                "failed"
            ],
            "x-enum-varnames": [
                "MessageStatusPending",
                "MessageStatusSent",
                "MessageStatusDelivered",
                "MessageStatusRead",
                "MessageStatusFailed"
            ]
        },
        "models.MessageType": {
            "type": "string",
            "enum": [
                "text",
                "image",
                "video",
                "audio",
                "document",
                "sticker",
                "template",
                "reaction"
            ],
            "x-enum-varnames": [
                "MessageTypeText",
                "MessageTypeImage",
                "MessageTypeVideo",
                "MessageTypeAudio",
                "MessageTypeDocument",
                "MessageTypeSticker",
                "MessageTypeTemplate",
                "MessageTypeReaction"
            ]
        },
...
        "models.SenderType": {
            "description": "who sent the message",
            "type": "string",
            "enum": [
                "contact",
                "agent",
                "system",
                "ai",
                "note"
            ],
            "x-enum-varnames": [
                "SenderContact",
                "SenderAgent",
                "SenderSystem",
                "SenderAI",
                "SenderNote"
            ]
        },
        "models.UpdateContactInput": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
...
        "models.UserPublic": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/models.UserRole"
                },
                "status": {
                    "$ref": "#/definitions/models.UserStatus"
                }
            }
        },
        "models.UserRole": {
            "type": "string",
            "enum": [
                "admin",
                "supervisor",
                "agent"
            ],
            "x-enum-varnames": [
                "RoleAdmin",
                "RoleSupervisor",
                "RoleAgent"
            ]
        },
        "models.UserStatus": {
            "type": "string",
            "enum": [
                "online",
                "away",
                "offline"
            ],
            "x-enum-varnames": [
                "StatusOnline",
                "StatusAway",
                "StatusOffline"
            ]
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}